Time: 1 ms (67.54%), Space: 51.5 MB (20.46%) - LeetHub

+++ 268-missing-number/268-missing-number.java
@@ -1,10 +1,10 @@
 class Solution {
     public int missingNumber(int[] nums) {
-        int n= nums.length;
-        long sum = (n*(n+1))/2;
-        for(int i=0;i<n;i++){
+        int n = nums.length;
+        int sum = n*(n+1) /2;
+        for(int i=0; i<n; i++) {
             sum -= nums[i];
         }
-    return (int) sum;
-  }
+        return sum;
+    }
 }
